Add shared schema contract feasibility study (NO-GO)
Produces docs/plans/SHARED_SCHEMA_CONTRACT.md, a go/no-go investigation
into whether the FlowTree controller API and ar-manager MCP server can
derive from a single shared OpenAPI-style spec to eliminate the dual-copy
schema divergence problem.

Verdict: NO-GO. The two-copy problem relocates rather than disappears:
- The controller validates fields via JsonFieldExtractor calls in Java
  with no mechanism to derive from an external spec without a full redesign.
- FastMCP from_openapi() belongs to the standalone fastmcp library, not the
  mcp>=1.0.0 package this codebase uses — switching requires rewriting the
  6,510-line server.py.
- MCP docstrings contain context-specific multi-sentence descriptions that
  have no auto-generation source; they require human authoring regardless.
- Legitimate surface differences (internal-only fields, MCP-only params)
  would require per-field curation markers — reintroducing the exact manual
  bookkeeping the approach was meant to eliminate.

The existing McpToolWorkstreamConfigSurfaceTest.java already provides
CI-enforced divergence detection. Extending that test is the recommended
path forward.
