+++ flowtree/runtime/src/main/java/io/flowtree/api/FlowTreeApiEndpoint.java
@@ -59,6 +59,8 @@ import io.flowtree.slack.SlackNotifier;
 import io.flowtree.slack.NotifierRegistry;
 import io.flowtree.workstream.WorkstreamConfig;
 import io.flowtree.submission.PhaseConfigResolver;
+import io.flowtree.workstream.WorkspaceEntry;
+import io.flowtree.workstream.WorkspaceSecretEntry;
 
 /**
  * HTTP API endpoint for FlowTree orchestration.
@@ -128,7 +130,7 @@ public class FlowTreeApiEndpoint extends NanoHTTPD implements ConsoleFeatures {
     private Map<String, String> orgToWorkspaceId = new HashMap<>();
 
     /** Resolves a Slack workspace ID to its entry for the workspace runner layer. */
-    private Function<String, WorkstreamConfig.WorkspaceEntry> workspaceLookup = id -> null;
+    private Function<String, WorkspaceEntry> workspaceLookup = id -> null;
     /** Tracks which jobs should have a PR auto-created on success. */
     private final Map<String, AutoPrContext> autoCreatePrJobs = new HashMap<>();
 
@@ -274,7 +276,7 @@ public class FlowTreeApiEndpoint extends NanoHTTPD implements ConsoleFeatures {
     }
 
     /** Sets the workspace lookup feeding the workspace layer of {@link SubmissionConfigResolver}; {@code null} disables it. */
-    public void setWorkspaceLookup(Function<String, WorkstreamConfig.WorkspaceEntry> lookup) {
+    public void setWorkspaceLookup(Function<String, WorkspaceEntry> lookup) {
         this.workspaceLookup = lookup != null ? lookup : id -> null;
     }
 
@@ -285,9 +287,9 @@ public class FlowTreeApiEndpoint extends NanoHTTPD implements ConsoleFeatures {
      * the workspace for a listener workstream and propagate any
      * workspace-level routing defaults.
      */
-    public WorkstreamConfig.WorkspaceEntry workspaceLookupOrNull(String workspaceId) {
+    public WorkspaceEntry workspaceLookupOrNull(String workspaceId) {
         if (workspaceId == null || workspaceId.isEmpty()) return null;
-        WorkstreamConfig.WorkspaceEntry entry = workspaceLookup.apply(workspaceId);
+        WorkspaceEntry entry = workspaceLookup.apply(workspaceId);
         return entry;
     }
 
@@ -364,7 +366,7 @@ public class FlowTreeApiEndpoint extends NanoHTTPD implements ConsoleFeatures {
      * @param cache workspace-ID → (secret-name → entry) map
      */
     public void setSecretsCache(
-            Map<String, Map<String, WorkstreamConfig.WorkspaceSecretEntry>> cache) {
+            Map<String, Map<String, WorkspaceSecretEntry>> cache) {
         secretsHandler.setSecretsCache(cache);
     }
 
@@ -849,7 +851,7 @@ public class FlowTreeApiEndpoint extends NanoHTTPD implements ConsoleFeatures {
         // resolver. Same code path the Slack listener uses, so workspace-level
         // settings are honoured uniformly regardless of submission entrypoint.
         String wsId = workstream.getWorkspaceId();
-        WorkstreamConfig.WorkspaceEntry wsEntry = (wsId != null && !wsId.isEmpty()) ? workspaceLookup.apply(wsId) : null;
+        WorkspaceEntry wsEntry = (wsId != null && !wsId.isEmpty()) ? workspaceLookup.apply(wsId) : null;
         if (wsId != null && !wsId.isEmpty() && wsEntry == null) log("submitWorkspaceMissing workspaceId=" + wsId);
         PhaseConfigBundle requestBundle = PhaseConfigResolver.bundleFromRequest(body);
         SubmissionConfigResolver configResolver = SubmissionConfigResolver.resolve(
