+++ flowtree/runtime/src/test/java/io/flowtree/slack/SlackMultiTenantConfigTest.java
@@ -31,6 +31,8 @@ import java.util.List;
 import java.util.Map;
 
 import static org.junit.Assert.*;
+import io.flowtree.workstream.WorkspaceEntry;
+import io.flowtree.workstream.WorkstreamEntry;
 
 /**
  * Tests for multi-tenant Slack workspace configuration, channel-key routing,
@@ -43,7 +45,7 @@ import static org.junit.Assert.*;
  */
 public class SlackMultiTenantConfigTest extends TestSuiteBase {
 
-    /** WorkstreamConfig.WorkspaceEntry serializes and deserializes all Slack fields. */
+    /** WorkspaceEntry serializes and deserializes all Slack fields. */
     @Test(timeout = 10000)
     public void testWorkspaceEntryYamlRoundTrip() throws IOException {
         String yaml = "slackWorkspaces:\n" +
@@ -60,7 +62,7 @@ public class SlackMultiTenantConfigTest extends TestSuiteBase {
         assertNotNull(config.getSlackWorkspaces());
         assertEquals(1, config.getSlackWorkspaces().size());
 
-        WorkstreamConfig.WorkspaceEntry entry = config.getSlackWorkspaces().get(0);
+        WorkspaceEntry entry = config.getSlackWorkspaces().get(0);
         assertEquals("T0123456789", entry.getId());
         assertEquals("T0123456789", entry.getSlackTeamId());
         assertEquals("my-org", entry.getName());
@@ -90,14 +92,14 @@ public class SlackMultiTenantConfigTest extends TestSuiteBase {
 
         assertEquals(2, config.getSlackWorkspaces().size());
 
-        WorkstreamConfig.WorkspaceEntry ws1 = config.getSlackWorkspaces().get(0);
+        WorkspaceEntry ws1 = config.getSlackWorkspaces().get(0);
         assertEquals("T111", ws1.getId());
         assertEquals("workspace-one", ws1.getName());
         assertEquals("xoxb-one", ws1.getBotToken());
         assertNotNull(ws1.getGithubOrgs());
         assertEquals("ghp_one", ws1.getGithubOrgs().get("my-org").getToken());
 
-        WorkstreamConfig.WorkspaceEntry ws2 = config.getSlackWorkspaces().get(1);
+        WorkspaceEntry ws2 = config.getSlackWorkspaces().get(1);
         assertEquals("T222", ws2.getId());
         assertEquals("/config/slack-tokens.json", ws2.getTokensFile());
     }
@@ -146,11 +148,11 @@ public class SlackMultiTenantConfigTest extends TestSuiteBase {
 
         assertEquals(2, config.getWorkstreams().size());
 
-        WorkstreamConfig.WorkstreamEntry entryWithId = config.getWorkstreams().get(0);
+        WorkstreamEntry entryWithId = config.getWorkstreams().get(0);
         assertEquals("C001", entryWithId.getChannelId());
         assertEquals("T111", entryWithId.getWorkspaceId());
 
-        WorkstreamConfig.WorkstreamEntry entryWithoutId = config.getWorkstreams().get(1);
+        WorkstreamEntry entryWithoutId = config.getWorkstreams().get(1);
         assertEquals("C002", entryWithoutId.getChannelId());
         assertNull(entryWithoutId.getWorkspaceId());
 
@@ -162,7 +164,7 @@ public class SlackMultiTenantConfigTest extends TestSuiteBase {
     /** SlackTokens.from(entry) extracts inline botToken and appToken. */
     @Test(timeout = 10000)
     public void testSlackTokensFromEntryInlineTokens() throws IOException {
-        WorkstreamConfig.WorkspaceEntry entry = new WorkstreamConfig.WorkspaceEntry();
+        WorkspaceEntry entry = new WorkspaceEntry();
         entry.setId("T999");
         entry.setBotToken("xoxb-inline-bot");
         entry.setAppToken("xapp-inline-app");
@@ -181,7 +183,7 @@ public class SlackMultiTenantConfigTest extends TestSuiteBase {
         Files.write(tempFile.toPath(),
                 "{ \"botToken\": \"xoxb-from-file\", \"appToken\": \"xapp-from-file\" }".getBytes());
 
-        WorkstreamConfig.WorkspaceEntry entry = new WorkstreamConfig.WorkspaceEntry();
+        WorkspaceEntry entry = new WorkspaceEntry();
         entry.setId("T888");
         entry.setTokensFile(tempFile.getAbsolutePath());
         entry.setBotToken("xoxb-should-be-ignored");
